requirements.txt
streamlit==1.33.0
moviepy==1.0.3
openai-whisper==20231117
python-docx==0.8.11
PyPDF2==3.0.1
python-magic==0.4.27
pydub==0.25.1
ffmpeg-python==0.2.0